Moves ProductShelf to sections folder
