+++ src/components/sections/ProductShelf/index.tsx
