+++ src/components/sections/ProductShelf/product-shelf.scss
