+++ src/pages/index.tsx
@@ -6,8 +6,7 @@ import type { PageProps } from 'gatsby'
 import type { HomePageQueryQuery } from '@generated/graphql'
 import BannerText from 'src/components/sections/BannerText'
 import Hero from 'src/components/sections/Hero'
-
-import ProductShelf from '../components/product/ProductShelf'
+import ProductShelf from 'src/components/sections/ProductShelf'
 
 import '../styles/pages/index.scss'
